CRAS: Move audio thread snapshot buffer to the end of server state
This is a non-backward compatible change to the server state ABI. This
patch should land together with the ARC++ side ABI update.

The snapshot buffer contains debug structs like audio_dev_debug_info.
Having it in the middle of the server state makes it difficult to change
the debug structs due to broken ABI with the ARC++ side. This patch
moves it to the end with other debug structs.

BUG=b:250544745
TEST=Unittests and cras_test_client -dump_events on Redrix.

Change-Id: I4f650d709cd39c5cda8aef96f4a3ca021b85e2c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5120854
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Pattara Teerapong <pteerapong@chromium.org>
Commit-Queue: Ben Zhang <benzh@chromium.org>