ucmlint: override Enum.__str__() for PEP-0663
With regard to https://peps.python.org/pep-0663/ on PY3.11, the
string-typed Enum behaves in a different way. StrEnum was
introduced to the rescue accordingly.

This CL is adopted preferably as the alternative to StrEnum usage,
with consideration to better compatibility to former versions.

BUG=none
TEST=manual test

Change-Id: I63d02e6ed32241f4c7fac34b0e88f9ffe6905826
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5237835
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Auto-Submit: Pin-chih Lin <johnylin@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Pin-chih Lin <johnylin@chromium.org>