cras: fix remainder calculation in audio thread
The remainder should be reduced by the actual number of frames retrieved from the input buffer rather than the expected number of frames.

Due to nread representing the possible number of frames to be read from the buffer rather than the actual number of frames read, the original computation was inaccurate.

As a result, the reminder value diff in the AUDIO_THREAD_READ_AUDIO and AUDIO_THREAD_READ_AUDIO_DONE events was computed incorrectly. The diff should reflect the frames actually read from the buffer, not the frames expected to be read from the buffer.

BUG=b:324026284
TEST=deploy to brya and see the atlog

Change-Id: I486d5a83ecb2ff19155b8eea73a5570bbbd8b26a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5252639
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Ching Yun Chang <whalechang@google.com>
Commit-Queue: Ching Yun Chang <whalechang@google.com>