CRAS: disable MSBC if SCO unused and adapter does not support
There was a case where CRAS could read the UCM and decide not to use
SCO, but the adapter supports WBS only via SCO, which results in a
mismatched expectation on using WBS per host-peer negotiation.

In this CL, we disable the MSBC codec for that particular case.

BUG=b:321180937
TEST=Verify WBS works (or not) depending on the queried result

Cq-Depend: chromium:5334439
Change-Id: I6290c3d922b5347566f9c972c8cbe7919f97dd3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5312827
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>