Add library for optional thread annotations
Use thread-specific context structs to prevent calling
functions or accessing global variables on the wrong thread.
Functions that should only be called on a specific thread
takes the thread-specific struct as the first argument.
Global variables that should be only accessed from a
specific thread are stored on the context struct.
This allows us to catch misuses at compile time,
which is very early in the development cycle.

This design features a run-time fallback so it is
non-invasive and allows incremental adoption.
When a function isn't annotated yet but needs to access a
function or global variable that requires the use of the
thread contexts, it can use helper functions to assert
that it is running under the correct thread and acquire
the context. Code anywhere in the dependency graph can
start adoption immediately, without requiring all
dependencies or all reverse dependencies to switch over.

BUG=b:330875939
FIXED=b:330875939
TEST=CQ

Change-Id: Ie4fa6d2d089fd86e306a0d84e3b21dec2393f666
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5383121
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Tested-by: Li-Yu Yu <aaronyu@google.com>