CRAS: rework SCO reconnection logic
When we receive an unsolicited SCO disconnection event, we will
reconnect to the SCO to prevent the call from hanging.

Whether it is solicited was determined by whether a
|StopScoCall| DBus call was invoked before it. However, there are cases
where that flag manages to survive till the next connection, in which
case can disrupt the logic.

This CL reworks it so that at the beginning of each connection, the flag
is ensured to be cleared.

Once |StopScoCall| has been invoked by CRAS, we should ignore all
disconnection events (i.e., not to reconnect) until the next proper
connection.

BUG=b:327557025,b:327696878
TEST=verify SCO reconnects after consecutive drops

Change-Id: Id98adeef35efbb264a076abc61b7bea4401eed62
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5446837
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>