ucm: fix JieLi BR21 playback sound too small
JieLi BR21 incorrectly reports the volume range as 0,
resulting in trigger software volume.

Using software volume is not a problem.
The problem is the volume range on the device is incorrect,
resulting in the software volume not being able to set
to the maximum.

This fix ensures that the mixer control is set to the
maximum volume and also flags the use of software volume.

BUG=b:328817525
TEST=verify the fix with feedback report user

Change-Id: I4f85fb55b577ea10f00379a701b982b242a0b8f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5457176
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Ching Yun Chang <whalechang@google.com>