CRAS: free SR resources after SCO callback is removed
In CL:5449295 we added a mutex lock to ensure that the SCO callback does
not access SR resources while it might be in the process of being
removed in the main thread.

Specifically, we are afraid of the situtation where the input device is
closing while the SCO callback is running due to the output device being
alive.

This can be avoided by freeing the resources only when the last device
is closed, since that is the timing to ensure the offending callback had
been removed.

Note that the similar issue nearing the construction of the SR resources
is not there due to the inversed order of creating/accessing them.

BUG=b:333619168
TEST=test scenario 2 in bug and verify no crash

Change-Id: I241d9317c6809bcad885aa0416c77c552b8fef0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5461198
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>