cras_stream_apm: Move down possibly_track_voice_activity
I want to move webrtc_apm_process_stream_f into
CrasProcessor to run together with the rest of the
effects.

possibly_track_voice_activity is a function to get
metadata out of the APM, doesn't process
audio so it doesn't need to be in the middle.

BUG=b:261951580
TEST=tast run dut1 audio.CrasSpeakOnMute\*

Change-Id: I55229b67a343dbb631b961a4c541748f8f3d9aa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5493010
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>