cras_processor: Remove "success" return value
The semantics of cras_processor_create's
bool return "success" and the processor stored in "ret"
is confusing.

The caller is expected to handle all the combinations:
1.  success==true, ret!=NULL
    => created successfully as specified
2.  success==false, ret!=NULL
    => the processor is degraded
3.  success==false, ret==NULL
    => some bug happened in cras_processor_create

Case 2 is subtle, and it only exists because we wanted
to expose the degraded status so we can report it
to UMA and audio thread dumps.

This CL instead make cras_processor_create just
return the plugin_processor and the actual effect
directly to clear confusion.

BUG=b:261951580
TEST=tast run dut1 audio.CrasEffects.\*

Change-Id: Iaa30b74abbc2cb09cfc0215a1b848aa770475c49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5500019
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>