cras_stream_apm: Apply aecdump to future APM
When we enable aecdump via chrome://webrtc-internals, it is only applied
to the current active APM on that stream. The APM may be removed and a
new one is created, or the active APM can be changed due to iodev
switching. This cause the future APM to not have aecdump applied.

Store the aecdump state on the cras_stream_apm and apply it to the
future APM.

BUG=b:333982642
TEST=Manual test:
     1. Go to chrome://webrtc-internals and start aecdump
     2. Start Meet
     3. Stop aecdump
     4. Unpack the dump and verify that the last stream has
        aec_enabled=1 and the reverse.wav is not empty
TEST=Manual test (Switching device):
     1. Start aecdump
     2. Select internal mic
     3. Start Meet
     4. Switch input to 3.5mm mic jack
     5. Stop aecdump
     6. Check that there is a dump for each input, unpack and check the
        dump
TEST=tast run $DUT audio.ChromeAECDump

Change-Id: I4a807ba3a612d0a4ac19d5761aceebff0d2f7c89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5524641
Commit-Queue: Pattara Teerapong <pteerapong@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>