cras_stream_apm: Skip processing on the main thread
process_reverse might be unexpectedlly called from the
main thread, if an iodev does not have start ops and
cras_iodev_fill_odev_zeros is used as a fallback.

This CL workarounds the assertion failure by doing nothing
in process_reverse() if it finds itself called out of the
audio thread. Skipping this data for the reverse apm stream
is OK because the zeros are expected to not be helpful for
echo cancellation anyway.

BUG=b:340129658
TEST=tast run dut12 audio.CrasMultiChannelCapture.\* audio.CrasAEC.\*

Change-Id: Ic3cdf3f0b6a007e39f97d01e5f3d823171461653
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5534460
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>