floss: close socket after |stop_sco_call|
If the socket is closed before the SCO is stopped, there is a chance
that the BT stack will try to write to the socket with full but futile
effort that possibly breaks things.

This CL closes the socket after |stop_sco_call| is resolved.

BUG=b:333651832
TEST=switch profile multiple times

Change-Id: I6791dff9d2b5dfaff38d01a31e2d8999593b4a25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5547045
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>