floss: remove 500ms of delay in context switch
We added the delay due to unknown reason that would otherwise cause the
capture path to fail with recurrent noise.

This was found to be related to resetting the active group, and now we
keep the same group active for the device lifetime, the delay is no
longer required.

BUG=b:317682584
TEST=Verify context switch works flawlessly

Change-Id: I7835fe0a121ffd8039daf38fb7aa7585bbcbb65f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5584118
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>