bt_policy: block a2dp open when delayed cb is alive
When there is a profile switch from HFP to A2DP, there is an intentional
delay (500ms) for reopening the output device due to observed results.

However, that is not guarded well, where the output device can be opened
upon any request, and that can cause problems because it is deactivated
in this window, relying on the callback to reactivate and resume.

This CL guards |a2dp->open_dev| so that it cannot be opened while the
callback is active.

BUG=b:342943472
TEST=Verify the delay and that it opens successfully at first open

Change-Id: If0de4212582f224e1d5613da4e27e97585b23156
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5622437
Reviewed-by: Ching Yun Chang <whalechang@google.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>