floss: improve |hfp_write| strategy
|hfp_write| relies on the |POLLIN| event coming from Floss, which
usually asks for 1 or 2 packets of data.

1 packet of data is worthy of 7.5ms of data, which is different from the
block size being registered at CRAS, which is 10ms.

When HFP is opened, the input device opens first, then the output
device. Until the output is open, zeros are filled in to match the size
of the incoming stream. However, once the output is marked as started,
at that very moment it could be asked for 2 packets worth of data, which
it does not have.

In this CL, we change the block size to 7.5ms, and stuff 2 packets worth
of zero at the beginning for the output device.

BUG=b:345741433
TEST=Verify significantly fewer "poll timeout" in HFP

Change-Id: Ie264183a17842295347439e2b3c5e9a1843a5212
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5620376
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>