cras: dsp: Add testtool for comparing behaviors of C code and rust code
Use python to write a script which complie the dsp tests with libraries
in C and rust respectively, and run the two binaries with same input to
compare their behaavior. To run the test tool, an additional 'adhd'
directory is needed and would be passed to the tool as an argument.

BUG=None
TEST=None

Change-Id: Iabb49d86469bc0fff06cf4dc08b1fe1ea756fa08
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5677488
Commit-Queue: Nathan Lee <nathanmyl@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>