sound_card_init: Refactor for CalibData and VPD
CalibData and VPD previously worked on the assumption that all smart
amps just requires an r0 and temp value, and the initialization was
similarly hard coded.

Adjusted CalibData to use the VPDTrait interface instead, which only
requires the channel count to adapt to TAS2563 and other smart amps
better. VPD is adjusted to implement the VPDTrait interface.

BUG=b:346949244
TEST=tast run $DUT audio.SoundCardInit.*
TEST=sound_card_init debug --id sofrt5682 --amp $(cros_config /audio/main speaker-amp) --conf $(cros_config /audio/main sound-card-init-conf) on delbin and redrix

Change-Id: I73d6c3c904beea3c1b0ce6c68bf1dfa3e71bf343
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5743138
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Baili Deng <bailideng@google.com>
Reviewed-by: Judy Hsiao <judyhsiao@google.com>