Move CrasDlcId into types_internal.rs
This is mainly for s2 to access CrasDlcId and avoid cyclic dependency.
Before this change, dlc depends on s2, and CrasDlcId is defined in dlc.
After this change, dlc depends on types_internal and s2, and s2 can
depend on types_internal.

Per discussion, we decided not to make it a crate to keep the original
design: 1 crate with 1 header.

BUG=b:359803173
TEST=bazel test //...

Change-Id: I474fc294d5e6e0d6db4f7378c6b7ebd7d2ea41e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5790042
Tested-by: Cranel W <cranelw@chromium.org>
Commit-Queue: Cranel W <cranelw@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>