floss: add |is_sco_connected| logic
|is_sco_stopped| was introduced in CL:5446837 to distinguish if a
previous SCO disconnection was initiated by the host, which is a factor
of making the decision to reconnect the device.

However, when a disconnection is initiated by the host, and another
connection request is initiated by the host, and the disconnection
event arrives after then, CRAS has to determine that this was actually
caused by the previous disconnection request from the host and ignore it.

Here we introduce the strong assumption that, for every connected SCO,
there must be a disconnection event that entails. With this, we will
maintain a variable that allows CRAS to reject SCO open requests in the
event that the previous has not disconnected (i.e., fully respecting
the fact that |is_sco_stopped| doesn't necessarily imply
|!is_sco_connected| and vice versa).

BUG=b:352497401
TEST=Verify SCO works well with stress test

Change-Id: Ide640b8c19fbc3b802c4d1dfe0f25e919257de61
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5796322
Commit-Queue: Jeremy Wu <jrwu@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>