UCM: Add Lenovo docks:
1.ThinkPad Hybrid USB-C with USB-A Dock
2.ThinkPad USB-C Dock Audio

Disabling hardware volumn as default and enabling software volumn to fix the small playback sound.
This fix ensures that the mixer control is set to the maximum volume and flags the use of software volume.
Also specify the input & output audio node when connecting the dock.

BUG=b:357472584
TEST=Run commands
  (outside)$ cros workon --board nissa start adhd
  (outside)$ cros_sdk emerge-nissa media-sound/adhd
  (outside)$ cros deploy $IP media-sound/adhd
  (chromebook)$ restart cras

Change-Id: I6e0ec0dd4eb562359a87351dbe2efc561481f302
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5828311
Commit-Queue: Terry Cheong <htcheong@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Tested-by: Sarah Wang <sarah1.wang@lcfc.corp-partner.google.com>