floop: make no stream playback no-op
When playback is not active, flexible loopback should return frames
filled with zeros. But this is handled in input side since there would
be cases that the playback side didn't started at all.

Setting it to limit the source of filling zeros and avoid overfilling
the buffer.

BUG=b:362679648
TEST=tast run <dut> audio.FloopLifecycle.*

Change-Id: Ib8a0e158d994064de8a91f14c8c3149f3bc0e632
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5828312
Tested-by: Terry Cheong <htcheong@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Terry Cheong <htcheong@chromium.org>