CRAS: fix effect selection bug of Style Transfer and AP NC
On devices with Style Transfer, when internal mic is active,
the UI is a toggle for Style Transfer;
when external mic is active, the UI is a toggle for NC.

If
1. Select external mic.
2. Turn on the NC toggle.
3. Select internal mic.
4. Turn off the Style Transfer toggle.

The expected effect is NoEffects, but currently it will get NoiseCancellation.

This CL fixes the issue.

BUG=b:364812248
FIXED=b:364812248
TEST=bazel test //... --config=local-clang --//:apm

Change-Id: I7eaa07edb054e8965b4273f7fd138ce115eab657
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5839085
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Cranel W <cranelw@chromium.org>
Commit-Queue: Hung-Hsien Chen <hunghsienchen@chromium.org>