CRAS: Add system_valid_nc_providers in S2 Output
system_valid_nc_providers represents the NC providers that can be
used if the node is compatible and the effect should be enabled.
It will be used to decide the stream effect and the effect restart
tag of cras_iodev from S2.

BUG=b:353627012
TEST=bazel test //... --config=local-clang --//:apm

Change-Id: I618e84fed4d70293f296052e77de30f9c61de779
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5856749
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>