CRAS: Combine NC and ST enabled as voice_isolation_ui_enabled
There will be at most 1 toggle in UI setting for the voice isolation
effects.
The `enabled` flags of the effects including Style Transfer, AP/DSP NC,
and Beamforming are combined as 1 flag, voice_isolation_ui_enabled.

The bool noise_cancellation_enabled, which was used in
cras_system_state, is replaced and removed in this CL.

Added new D-Bus method SetVoiceIsolationUIEnabled.
Keeping SetNoiseCancellationEnabled and SetStyleTransferEnabled for now
because the UI side is not updated yet.

This CL requires changes to Tast in the meantime.

BUG=b:353627012
TEST=bazel test //... --config=local-clang --//:apm
TEST=CQ

Cq-Depend: chromium:5893135
Change-Id: I3f8105504cba3cd30bea83b2949c8c695d4999fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5873624
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Hung-Hsien Chen <hunghsienchen@chromium.org>