Add metrics for DLC installation failure
Before this change, we only log the number of retried times on success for dlc installation. However, if the installation keeps failing, there will be no metrics logged. This change sends the elapsed time on each failure.

Note that the logged unit is second, while the unit of metric on success is the number of retried times. This is for eliminating the burden to convert retried times to elapsed seconds.

BUG=b:367863686
TEST=bazel test //...

Change-Id: I92b37b293d1cd0797e9f092de66c28fb14900e92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5871020
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Tested-by: Cranel W <cranelw@chromium.org>
Commit-Queue: Cranel W <cranelw@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>