CRAS: Resolve CrasProcessorEffect for cras_stream_apm with S2
The effect was resolved in cras_processor_config.c.
Now that we have the effect factors in S2 as the source of truth, we
decided to move the effect resolving into S2.

DSP NC is set to not allowed when voice_isolation_ui_enabled is false.
This is for preventing DSP NC from being selected by client controlled
voice isolation.

The local states are logged in cras_stream_apm_add(), and the global
states in S2 are logged by audio_diagnostics.

BUG=b:353627012
TEST=bazel test //... --config=local-clang --//:apm

Disallow-Recycled-Builds: test-failures
Change-Id: I04b0a1afe40780d6f2a8baddb61a3ff7459d5da8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5881347
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>