ucm: force stereo playback on HyperX Cloud Revolver S
The bug linked below was investigated thoroughly and turned out to have
a complex and difficult to debug root cause, closely related to a particular
series of Intel chipset used in certain Chromebooks. It was decided that the
impact of the bug does not justify costly investigation of the root cause and
instead the only known affected device should have a workaround applied for the
issue.

The workaround is to forbid 8-channel playback on the affected device. This
commit implements it using UCM where 2-channel playback is set.

BUG=b/301014984
TEST=Simultaneous playback from Spotify and Meet call for >1h without
any instabilities

Change-Id: I780182746410e58443571fb44f86b8ebcf6525be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5889572
Reviewed-by: Ching Yun Chang <whalechang@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Jędrzej Ciupis <jciupis@google.com>