CRAS: Resolve iodev restart tag with S2
When open_dev, we will configure the iodev with a restart tag, which is
used to check if the audio effect has changed, and if so, the iodev
needs to be restarted.
Added cras_s2_get_iodev_restart_tag_nc_provider() to handle the
resolving from S2, including the restart tag and the new effect state
while effect factors change.

BUG=b:353627012
TEST=bazel test //... --config=local-clang --//:apm

Disallow-Recycled-Builds: test-failures
Change-Id: Ia85cc1bc0a59145ece06bc1841876acddb92069a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5901248
Commit-Queue: Hung-Hsien Chen <hunghsienchen@chromium.org>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>