cras_cbindgen: Generate code from crate root
Update the rule to take a rust_library. The library allows the rule
to know the entry point of the crate (crate_info.root) and all the
source files of the crate (crate_info.srcs).

This prevents duplicated declarations in generated headers.
The old implementation passes all srcs to to cbindgen and
I think it causes duplicated declarations because a function in
src/foo.rs may be generated for src/foo.rs as well as for
src/lib.rs's "pub mod foo" statement.

BUG=b:372902037
TEST=devtools/rust_generate.py
TEST=cargo build && SYSTEM_CRAS_RUST_LIB=$PWD/target/debug/libcras_rust.a bazel test --//:system_cras_rust //...
TEST=bazel test //...
TEST=FEATURES='test -usersandbox' emerge-amd64-generic cras_rust adhd

Change-Id: I0691e9cb58f2025d1f7167c5cc993130b9bf8480
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5920309
Reviewed-by: Cranel W <cranelw@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>