cras_cbindgen: Remove --add-keyword-*
The workaround for struct/enum keywords needs to specified
when a type is defined out of the current crate. However
that does not scale to chained dependencies.
For example: if we have the dependency chain
foo -> bar -> baz and we have types bar::T and baz::U,
we need separate flags to control whether keywords are
needed for T and U.

Instead of accurately control the keywords, let's just
make cbindgen generate the duplicated keywords and
workaround them.

BUG=b:372902037
TEST=devtools/rust_generate.py

Change-Id: I109c055aefc0e24f6cab8eb6da059630d385e0e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5920310
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Cranel W <cranelw@chromium.org>