sound_card_init: Add profiles and play zeros when applying TAS2563 Calibration Data
It is noticed that `Speaker Calibrated Data` was not written correctly
if the profile wasn't set correctly. Instead of the data written, it
will be using a default data or not be readable.

Set the profile to calibration and play zeros in order to write the
Speaker Calibrated Data properly.

BUG=b:383708600
TEST=`amixer -c 0 contents` and check that the Speaker Calibrated Data
is applied correctly
TEST=sound_card_init debug --id sofmt8188tas256 --amp TAS2563 --conf ciri.TAS2563.yaml

Cq-Depend: chromium:6109729
Change-Id: I929a2757994dfbc49748c98808a05bbad7d49917
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/6109127
Commit-Queue: Baili Deng <bailideng@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Judy Hsiao <judyhsiao@google.com>