bt_device: check if valid device before set_hw_vol
In b/146432097, a logic to reset some fields of the device on transport
removal was introduced.

However, we have found logs in the wild that shows the device can be
invalidated before removal of the transport that holds the device.

This CL is a low-risk attempt to at least validate the device before use
so it won't crash in such scenarios.

BUG=b:398082959
TEST=CQ

Change-Id: I6a40f47fcbadbc6426952b420354b649e640a776
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/6310149
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>