Update Absolute Capture Timestamp with internal device latency.
Some microphone introduced a internal latency (Logitech Rally bar) which
is unaccounted when timestamping the audio.

This introduce a UCM parameter for the node to be able to correct the
capture timestamp with the introduced latency by the device.

Those latencies are intrinsic to the devices and should only be device
dependant.

BUG=b:405377959
TEST=Not tested yet

Change-Id: Id98cc3885be38a028eeb88060ece8f3870911141
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/6402855
Commit-Queue: Lionel Koenig Gélas <lionelk@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Baili Deng <bailideng@google.com>