cras: Fix function pointer cast in hfp_slc_unittest
A UBSan runtime error was found in hfp_slc_unittest.cc due to a
call through a function pointer of an incorrect type.

The stub for cras_system_add_select_fd in the test defined a
callback with the signature void (*)(void*). However, the actual
function signature is void (*)(void* data, int revents). This
mismatch caused undefined behavior when the slc_cb callback was
invoked.

This change corrects the function pointer type for the callback,
updates the stub for cras_system_add_select_fd to match the real
signature, and modifies the calls to the callback to pass the
expected revents argument.

BUG=b:445167544
TEST=FEATURES='-usersandbox test' emerge-amd64-generic adhd

Change-Id: I5ff4c87aafce9e60df038325d0f59137e1e95796
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/6968557
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>