audio_processor: Add timeout to worker
The timeout is set to 10 seconds.
In the worker, the timeout is checked when "creating the pipeline" and
"processing the audio data with the pipeline".
If it times out, the worker will be terminated with SIGALRM.

The timeout mechanism uses alarm syscall.
For seccomp policy, added `alarm` for amd64, and
added `setitimer` for arm/arm64.

Added tests to ensure the worker crash is handled gracefully.
The timeout is set to 1 second for tests.

BUG=b:446725446
TEST=bazel test //audio_processor:integration_tests_tests/peer
TEST=bazel test --config=local-clang //...

Change-Id: I024a8121a3e43252db63432c71dd653a49a87985
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/7206655
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>